refactor: separate routers as middlewares (#42)
For separation of concerns.
