+++ routers/index.js
@@ -1,0 +1,8 @@
+exports.rootRouter = require('./root');
+exports.leaderboardRouter = require('./leaderboard');
+exports.aboutRouter = require('./about');
+exports.signinRouter = require('./signin');
+exports.signoutRouter = require('./signout');
+exports.answerRouter = require('./answer');
+exports.dayRouter = require('./day');
+exports.oauthRouter = require('./oauth');
